+++ carpetmodSrc/carpet/commands/CommandCounter.java
@@ -1,106 +1,147 @@
 package carpet.commands;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.ArrayList;
-import java.util.Locale;
-
-import javax.annotation.Nullable;
-
 import carpet.CarpetSettings;
 import carpet.helpers.HopperCounter;
+import carpet.helpers.ItemWithMeta;
 import carpet.utils.Messenger;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
+import net.minecraft.command.NumberInvalidException;
 import net.minecraft.command.WrongUsageException;
+import net.minecraft.item.EnumDyeColor;
+import net.minecraft.item.Item;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
 
-import net.minecraft.item.EnumDyeColor;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Locale;
 
-public class CommandCounter extends CommandCarpetBase
-{
+public class CommandCounter extends CommandCarpetBase {
+    @Nullable
+    @ParametersAreNonnullByDefault
+    public static ItemWithMeta parseItemWithMeta(ICommandSender sender, String[] args, int ordinal) throws NumberInvalidException {
+        if (ordinal >= args.length) {
+            return null;
+        }
+        Item item = getItemByText(sender, args[ordinal]);
+        ++ordinal;
+        int metadata = ordinal < args.length ? parseInt(args[ordinal]) : 0;
+        return new ItemWithMeta(item, metadata);
+    }
+
     /**
      * Gets the name of the command
      */
-    public String getUsage(ICommandSender sender)
-    {
-        return "Usage: counter <color> <reset/realtime>";
+    @Nonnull
+    @ParametersAreNonnullByDefault
+    public String getUsage(ICommandSender sender) {
+        return "Usage: /counter [<color>|cactus|all] (reset|stop|realtime|raw)";
     }
 
-    public String getName()
-    {
+    @Nonnull
+    public String getName() {
         return "counter";
     }
 
     /**
      * Callback for when the command is executed
      */
-    public void execute(MinecraftServer server, ICommandSender sender, String[] args) throws CommandException
-    {
-        if (CarpetSettings.hopperCounters == CarpetSettings.HopperCounters.off && !CarpetSettings.cactusCounter){
+    @ParametersAreNonnullByDefault
+    public void execute(MinecraftServer server, ICommandSender sender, String[] args) throws CommandException {
+        if (CarpetSettings.hopperCounters == CarpetSettings.HopperCounters.off && !CarpetSettings.cactusCounter) {
             msg(sender, Messenger.m(null, "Need cactusCounter or hopperCounters to be enabled to use this command."));
             return;
         }
         if (args.length == 0) {
-            msg(sender, HopperCounter.formatAll(server, false));
+            Messenger.send(sender, HopperCounter.formatAll(false, true));
             return;
         }
         switch (args[0].toLowerCase(Locale.ROOT)) {
+            case "raw":
+                Messenger.send(sender, HopperCounter.formatAll(false, false));
+                return;
             case "realtime":
-                msg(sender, HopperCounter.formatAll(server, true));
+                Messenger.send(sender, HopperCounter.formatAll(true, false));
                 return;
             case "reset":
-                HopperCounter.resetAll(server);
-                notifyCommandListener(sender, this, "All counters restarted.");
+                HopperCounter.resetAll(true);
+                notifyCommandListener(sender, this, "All counters restarted instantly.");
+                return;
+            case "stop":
+                HopperCounter.resetAll(false);
+                notifyCommandListener(sender, this,
+                        "All counters have stopped and will restart on triggered.");
+                return;
+            case "distribution":
+                Messenger.send(sender, HopperCounter.formatAllDistribution(parseItemWithMeta(sender, args, 1)));
                 return;
         }
         HopperCounter counter = HopperCounter.getCounter(args[0]);
         if (counter == null) throw new WrongUsageException("Invalid color");
         if (args.length == 1) {
-            msg(sender, counter.format(server,false, false));
+            Messenger.send(sender, counter.format(false, false, true));
             return;
         }
         switch (args[1].toLowerCase(Locale.ROOT)) {
+            case "raw":
+                Messenger.send(sender, counter.format(false, false, false));
+                return;
             case "realtime":
-                msg(sender, counter.format(server, true, false));
+                Messenger.send(sender, counter.format(true, false, false));
                 return;
             case "reset":
-                counter.reset(server);
-                notifyCommandListener(sender, this, String.format("%s counters restarted.", args[0]));
+                counter.reset(true);
+                notifyCommandListener(sender, this, String.format("%s counter restarted instantly.", args[0]));
+                return;
+            case "stop":
+                counter.reset(false);
+                notifyCommandListener(sender, this, String.format(
+                        "%s counter has stopped and will restart on triggered.", args[0]));
+                return;
+            case "distribution":
+                Messenger.send(sender, counter.formatDistribution(parseItemWithMeta(sender, args, 2)));
                 return;
         }
         throw new WrongUsageException(getUsage(sender));
 
     }
 
-    public List<String> getTabCompletions(MinecraftServer server, ICommandSender sender, String[] args, @Nullable BlockPos pos)
-    {
-        if (CarpetSettings.hopperCounters == CarpetSettings.HopperCounters.off && !CarpetSettings.cactusCounter)
-        {
+    @Nonnull
+    @ParametersAreNonnullByDefault
+    public List<String> getTabCompletions(MinecraftServer server, ICommandSender sender, String[] args, @Nullable BlockPos pos) {
+        if (CarpetSettings.hopperCounters == CarpetSettings.HopperCounters.off && !CarpetSettings.cactusCounter) {
             msg(sender, Messenger.m(null, "Need cactusCounter or hopperCounters to be enabled to use this command."));
-            return Collections.<String>emptyList();
+            return Collections.emptyList();
         }
-        if (args.length == 1)
-        {
-            List<String> lst = new ArrayList<String>();
+        if (args.length == 1) {
+            List<String> lst = new ArrayList<>();
             lst.add("reset");
-            for (EnumDyeColor clr : EnumDyeColor.values())
-            {
+            for (EnumDyeColor clr : EnumDyeColor.values()) {
                 lst.add(clr.name().toLowerCase(Locale.ROOT));
             }
             lst.add("cactus");
             lst.add("all");
             lst.add("realtime");
+            lst.add("stop");
+            lst.add("raw");
+            lst.add("distribution");
             String[] stockArr = new String[lst.size()];
             stockArr = lst.toArray(stockArr);
             return getListOfStringsMatchingLastWord(args, stockArr);
         }
-        if (args.length == 2)
-        {
-            return getListOfStringsMatchingLastWord(args, "reset", "realtime");
+        if (args.length >= 2) {
+            if ("distribution".equalsIgnoreCase(args[args.length - 2])) {
+                return getListOfStringsMatchingLastWord(args, Item.REGISTRY.getKeys());
+            }
+            if (args.length == 2) {
+                return getListOfStringsMatchingLastWord(args, "reset", "realtime", "stop", "raw", "distribution");
+            }
         }
-        return Collections.<String>emptyList();
+        return Collections.emptyList();
     }
 }